Fix glVertexAttribPointer for client-side arrays
We need to store a copy to the data :(
PyOpenGL does this too. I think the problem is that the data is not
actually read from the pointer until draw time. In contrast to e.g.
glBufferData or glTexImage2D where the data is uploaded at the moment
of the call.