Change how gl.pyopengl imports its functions from OpenGL.GL
This was done using an explicit import, but some implementations
(like some machines on Travis) do not have all names. Switched to
an approach similar to what we have before: _pyopengl defines a list
of function names to import. in pyopengl.py we dynamically inject
these names, and wrap them into a dummy function if it is not available.