Create a BaseGLProxy class in gloo/gl/_proxy.py
This class is autogenerated and can be subclassed for different
purposes, such as a main proxy, a debug proxy, webgl or IPython
backend, etc.

By putting the functions in a class, a new set of functions can
be easily "created" simply by instantiating the class. The functions
all use __call__ to "serialize" the API calls. Subclasses can thus
simply implement __call__ in the appropriate way.