Fix loading of OpenGL lib on OSX.
Also raise error if the lib could not be found. Previously
the lib was not found on osx, resulting ctypes in loading
a null-lib or something, which acted as a sort of proxy
for *any* loaded lib. This error was undetected for quite a while
because the opengl lib was loaded by PyOpenGL or an app backend.
See #201