app: pyside and pyqt4 are now more normal backends. qt is no backend.
This makes it that qt/pyside/pyqt4 has less of a special status, except
for that most code for both backends is shared.

It is now actually possible to use both backends at the same time.
Previously it seemed that this was possible, but you'd get whatever
was selected the first time that _qt was loaded. Of course, it not
recommended to use both at the same time, but our tests seem to need
it, and so does sys_info().

Also, when auto-selecting a backend, vispy prefers backends for which the
GUI toolkit is already loaded (this was already the case). If such a
backend could *not* be used, a warning is issued.