Add example for multi-texturing
I added this while trying to find the bug in the sdf test. Funny thing
is that the error was that texture handles were used where I should have
used texture units. When there is only one program involved, these
are equal. Therefore this example did not reveal the bug. Still nice
to have though ...