parent aa9c01dae0c45afae6d177c1ab014c9e34bc6fc0 author Paul Debus <paul.debus@uni-weimar.de> 1534870849 +0200 committer Paul Debus <paul.debus@uni-weimar.de> 1562743668 +0200
Support drawing only edges of the mesh

To only draw the edges of a mesh, or to produce a compound visual that
displays the edges, the "lines" mode of the mesh can be used. This
requires passing in the faces not as triangles but as edge list, so not
as an (n_faces, 3) array, but (n_edges, 2) array. This commit also
allows for passing in the face index matrix, that will be converted into
an edge index matrix. Should work for all meshes with constant type
polygons, so triangle or quad meshes.