Add total_rotation_angle and total_scale_factor to touch (pinch) events.
Those two are supposed to hold the total scale change and rotation
change since the beginning of the touch event, and are easier to
manipulate than to accumulate the delta changes.

Note that testing locally on MacOS 10.12 using a macbook pro with a
touchpad the rotation does not seem to follow the description of the Qt
framework. That is to say:

 - event.rotation: total rotation since begin event
 - event.scale: Change of scale since last event, generally between 0.95 and 1.05.
 - event.total_rotation_angle: Always 0 for me
 - event.total_scale_factor: Total change of scale change since evt.type=begin event.

I still include those two, as in particular it seem on some devices you
can "interrupt" and "restart" a gesture; especially for rotation, and
even if total_rotation_angle seem broken on my end; total_scale_factor is
useful.

Closes #1964