Remove TextVisual optimization that ignores STTransform translation
The 'text_scale' STTransform used in the TextVisual never has its
translate property set, but the shader is also setting pos.w to 0. This
causes an "optimization" in some GPUs/compilers to remove the translate
uniform altogether causing VisPy to attempt to set an inactive uniform
later on (translate defaults to 0). This commit sets pos.w to 1.0 which
should have the same effect (no translation) while still preserving the
uniform.