Various optimizations related to camera linking and mesh normals (#2532)
See #2530 

CC @jintaoLee-roger

This is a collection of optimizations uncovered in #2530 debugging by me
or @jintaolee-roger. The main ideas are to not update camera properties
unless the value has actually changed. This prevents expensive
transform-updating methods from being called unnecessarily. This is
especially noticeable when a single camera move updates 4+ properties of
a linked property and each one tries to report that the view/transform
has changed. It does this in two ways:

1. Don't execute a transform update if we're updating properties
2. Don't update a property (and signal a view change) if the property
value is the same

This PR also adds some checks in the mesh light filter to see if vertex
normals have actually changed before submitting/uploading the normals to
the GPU. It also enables the use of a previously unused (yet defined for
some reason?) `MeshData._vertex_normal_indexed_by_faces`. This allows
the check in the light filter to actually work with a simple identity
(`is`) check.


TODO:

* Testing of performance by @jintaolee-roger
* Adding tests for at least the normals caching in the meshdata.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>