Skip conda pyopengl 3.1.9 (#2668)
Continuing issue from #2636.

For context: there's currently some issue with the conda-forge package
for pyopengl. It's causing several problems (not just our CI failures).
We are in the process of figuring out the issue, but for the time being
we just blacklist these versions here. This only affects CI and it
allows us to actually cut a release.