implementation spacing parameter grid mode (#2671)
In `Grid` on main, one can specify a spacing parameter, but it would not
have any affect. This PR finishes the implementation of the spacing
parameter. In addition, the spacing parameter can also now take a tuple
in order to define both `width_spacing` and `height_spacing`. The
docstring has been updated to reflect this.

Some additional notes regarding spacing. If spacing is defined, then
spacing will always be added in between sequential grid positions, no
matter whether the grid position is occupied by a `ViewBox` or not.

Lastly, the default spacing was set to 6. This has been changed to 0.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>