Bug fix: Incorrectly passed flag to re.sub (#2681)
The code uses `re.M` as a positional argument instead of as a keyword
argument in the `re.sub()` function. This is incorrect because
`re.sub()` expects the fourth argument to be count (the maximum number
of pattern occurrences to replace), not the flags.

The bug was found due to the deprecation warning in Python 3.13:
```
vispy/vispy/gloo/program.py:248: DeprecationWarning: 'count' is passed as positional argument!
```