Keep hold of text vertex data on cpu (#2684)
Currently it's impossible to get any info about the *rendered*
dimensions of the text. This is relevant for things like
https://github.com/napari/napari/pull/7836, where we need to know how
*big* the rendered text is on the cpu.

In order to allow downstream users to infer things like this, it pretty
easy for us to simply save the original vertex buffer on the CPU. I
hesitate to expose publicly, but at least this will unblock the feature
above.