Fix mouse buttons state handling
Under some circumstances a mouse press is not followed by a corresponding mouse
release event. The current implementation of recording which mouse buttons are pressed
relied on this behaviour. Luckily most backends record this information for us, so we
can use those functions instead of doing the bookkeeping ourselves. Backends which have
been fixed are:
- Qt
- jupyter_rfb
- tkinter
- wx

For the other backends I couldn't figure out how to do this as they do not provide
a simple way of fetching the mouse button state. Therefore the behaviour of the buttons
attribute is unchanged for the following backends:
- glfw
- pyglet
- sdl2