discard marker fragments if size was zero (#2696)
Noticed when working on https://github.com/napari/napari/pull/8223.

In some hardware (linux + nvidia is at least one), setting the size of
markers to `0` results in a value of `inf` for the border width due to a
division by zero.
This can have crazy results like the above link. In general, if the
point size is 0 we should simply not render it, I think, so this PR
simply discards the affected fragments.