Expose modifiers in SceneMouseEvent API (#2699)
Hi!

For my use case I would like to know whether shift was pressed when a
`SceneMouseEvent` happened, but the docs gave me no indication of such a
function. By reading the source code I found out that I could grab them
via the `mouse_event` attribute, but I figure it's not supposed to be
accessed directly because it's not marked as a property. This PR thus
exposes the modifiers as a property of `SceneMouseEvent`.